requirements.txt
notion-client==2.3.0
python-dotenv==1.0.1
openai==1.63.2
